--- a/src/0install-solver/hash_set.mli
+++ b/src/0install-solver/hash_set.mli
@@ -1,0 +1,7 @@
+type t
+
+val create : unit -> t
+val is_empty : t -> bool
+val add : t -> int -> unit
+val mem : t -> int -> bool
+val clear : t -> unit