main:
```
Benchmark 1: ~/github/ocaml/dune/_build/default/bin/main.exe pkg lock
  Time (mean ± σ):      2.140 s ±  0.050 s    [User: 1.384 s, System: 0.501 s]
  Range (min … max):    2.078 s …  2.245 s    10 runs
```

PR:
```
Benchmark 1: ~/github/ocaml/dune/_build/default/bin/main.exe pkg lock
  Time (mean ± σ):      2.127 s ±  0.121 s    [User: 1.345 s, System: 0.490 s]
  Range (min … max):    2.055 s …  2.466 s    10 runs
```